[depthMap] CUDA: Rename file LRUCache
